refactor(api): GitHubControllerBase mit TryResolveOwnerRepo extrahieren (#187)
TryResolveOwnerRepo aus 4 Controllern in gemeinsame Basisklasse
GitHubControllerBase verschoben (~112 Zeilen Duplikat entfernt).
