+++ src/ghGPT.Api/Controllers/DiscussionsController.cs
@@ -8,7 +8,7 @@ namespace ghGPT.Api.Controllers;
 [Route("api/repos/{id}/discussions")]
 public class DiscussionsController(
     IRepositoryService repositoryService,
-    IDiscussionService discussionService) : ControllerBase
+    IDiscussionService discussionService) : GitHubControllerBase(repositoryService)
 {
     [HttpGet]
     public async Task<ActionResult<IReadOnlyList<DiscussionItem>>> GetDiscussions(string id, [FromQuery] int limit = 30)
@@ -43,36 +43,6 @@ public class DiscussionsController(
         catch (InvalidOperationException ex)
         {
             return BadRequest(new { error = ex.Message });
-        }
-    }
-
-    private bool TryResolveOwnerRepo(string repoId, out (string owner, string repo) result, out ActionResult? error)
-    {
-        result = default;
-        var repo = repositoryService.GetAll().FirstOrDefault(r => r.Id == repoId);
-        if (repo is null)
-        {
-            error = NotFound(new { error = "Repository nicht gefunden." });
-            return false;
-        }
-
-        if (string.IsNullOrEmpty(repo.RemoteUrl))
-        {
-            error = BadRequest(new { error = "Dieses Repository hat keinen GitHub-Remote." });
-            return false;
-        }
-
-        try
-        {
-            var parsed = RemoteUrlParser.Parse(repo.RemoteUrl);
-            result = (parsed.Owner, parsed.Repo);
-            error = null;
-            return true;
-        }
-        catch (InvalidOperationException ex)
-        {
-            error = BadRequest(new { error = ex.Message });
-            return false;
         }
     }
 }
