+++ src/ghGPT.Api/Controllers/PullRequestsController.cs
@@ -8,7 +8,7 @@ namespace ghGPT.Api.Controllers;
 [Route("api/repos/{id}/pull-requests")]
 public class PullRequestsController(
     IRepositoryService repositoryService,
-    IPullRequestService pullRequestService) : ControllerBase
+    IPullRequestService pullRequestService) : GitHubControllerBase(repositoryService)
 {
     [HttpGet]
     public async Task<ActionResult<IReadOnlyList<PullRequestListItem>>> GetPullRequests(
@@ -166,36 +166,6 @@ public class PullRequestsController(
         catch (InvalidOperationException ex)
         {
             return BadRequest(new { error = ex.Message });
-        }
-    }
-
-    private bool TryResolveOwnerRepo(string repoId, out (string owner, string repo) result, out ActionResult? error)
-    {
-        result = default;
-        var repo = repositoryService.GetAll().FirstOrDefault(r => r.Id == repoId);
-        if (repo is null)
-        {
-            error = NotFound(new { error = "Repository nicht gefunden." });
-            return false;
-        }
-
-        if (string.IsNullOrEmpty(repo.RemoteUrl))
-        {
-            error = BadRequest(new { error = "Dieses Repository hat keinen GitHub-Remote." });
-            return false;
-        }
-
-        try
-        {
-            var parsed = RemoteUrlParser.Parse(repo.RemoteUrl);
-            result = (parsed.Owner, parsed.Repo);
-            error = null;
-            return true;
-        }
-        catch (InvalidOperationException ex)
-        {
-            error = BadRequest(new { error = ex.Message });
-            return false;
         }
     }
 }
